@@ -1,4 +1,4 @@
-class UserData < ActiveRecord::Migration
+class UserData < ActiveRecord::Migration[4.2]
   def up
     add_column :users, :zip,          :string
     add_column :users, :phone_number, :string
