@@ -1,4 +1,4 @@
-class CreateRepoSubscriptions < ActiveRecord::Migration
+class CreateRepoSubscriptions < ActiveRecord::Migration[4.2]
   def change
     create_table :repo_subscriptions do |t|
       t.string      :user_name
