@@ -1,4 +1,4 @@
-class AddGithubToUsers < ActiveRecord::Migration
+class AddGithubToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :github, :string
     add_column :users, :github_access_token, :string
