@@ -1,4 +1,4 @@
-class AddReposToIssues < ActiveRecord::Migration
+class AddReposToIssues < ActiveRecord::Migration[4.2]
   def change
     add_column :issues, :repo_id, :integer
   end
