@@ -1,4 +1,4 @@
-class AddRepoSubToIssueAssignment < ActiveRecord::Migration
+class AddRepoSubToIssueAssignment < ActiveRecord::Migration[4.2]
   def change
     add_column :issue_assignments, :repo_subscription_id, :integer
   end
