@@ -1,4 +1,4 @@
-class AddAdminToUsers < ActiveRecord::Migration
+class AddAdminToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :admin, :boolean
   end
