@@ -1,4 +1,4 @@
-class AddFavoriteLanguagesToUser < ActiveRecord::Migration
+class AddFavoriteLanguagesToUser < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :favorite_languages, :string, array: true
   end
