@@ -1,4 +1,4 @@
-class AddDailyIssueLimitToUser < ActiveRecord::Migration
+class AddDailyIssueLimitToUser < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :daily_issue_limit, :integer
   end
