@@ -1,4 +1,4 @@
-class AddDelieveredToIssueAssignment < ActiveRecord::Migration
+class AddDelieveredToIssueAssignment < ActiveRecord::Migration[4.2]
   def change
     add_column :issue_assignments, :delivered, :boolean, {default: false}
     # add_index  :users, :github, unique: true
