@@ -1,4 +1,4 @@
-class StoreGithubRequestErrors < ActiveRecord::Migration
+class StoreGithubRequestErrors < ActiveRecord::Migration[4.2]
   def change
     add_column :repos, :github_error_msg, :text
   end
