@@ -1,4 +1,4 @@
-class RemoveUserIdFromIssueAssignments < ActiveRecord::Migration
+class RemoveUserIdFromIssueAssignments < ActiveRecord::Migration[4.2]
   def change
     remove_column :issue_assignments, :user_id
   end
