@@ -1,4 +1,4 @@
-class RemoveRepoNameFromRepoSubscriptions < ActiveRecord::Migration
+class RemoveRepoNameFromRepoSubscriptions < ActiveRecord::Migration[4.2]
   def change
     remove_column :repo_subscriptions, :repo_name
   end
