@@ -1,4 +1,4 @@
-class RemoveUserNameFromRepoSubscriptions < ActiveRecord::Migration
+class RemoveUserNameFromRepoSubscriptions < ActiveRecord::Migration[4.2]
   def change
     remove_column :repo_subscriptions, :user_name
   end
