@@ -1,4 +1,4 @@
-class CreateDataDumps < ActiveRecord::Migration
+class CreateDataDumps < ActiveRecord::Migration[4.2]
   def change
     create_table :data_dumps do |t|
       t.text :data
