@@ -1,7 +1,7 @@
 require 'test_helper'
 
 class Users::OmniauthCallbacksControllerTest < ActionController::TestCase
-  include Devise::TestHelpers
+  include Devise::Test::ControllerHelpers
 
   setup do
     request.env["devise.mapping"] = Devise.mappings[:user]
